+++ src/main/java/com/evilnotch/lanessentials/commands/CommandNick.java
@@ -1,6 +1,6 @@
 package com.evilnotch.lanessentials.commands;
 
-import com.evilnotch.lanessentials.api.CapHandler;
+import com.evilnotch.lanessentials.cap.CapHandler;
 
 import net.minecraft.command.CommandBase;
 import net.minecraft.command.CommandException;
